Upper level keymaster code; fix compile issue and instantiate
This is the proper code to be used for the upper layer
keymaster code.

All builds and instantiations of this upper layer
use a commone lower level libary name for keymaster; we no longer
use the platform specific name for the lower level
library requirement.

Selection of the correct lower
level library is based on the selection of what is
being build by using ifdef in the Android.mk file
in each platform's lower level library so that only
the proper one is built.

Change-Id: I7889ffc98f18513c27ce72e78f13181ec9929107
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-285
Signed-off-by: Mark Allyn <mark.a.allyn@intel.com>
Reviewed-on: https://android.intel.com:443/247007